refactor(fiatconnect): move client creation outside quote class
