+++ src/fiatExchanges/quotes/FiatConnectQuote.ts
@@ -1,3 +1,4 @@
+import { FiatConnectApiClient } from '@fiatconnect/fiatconnect-sdk'
 import {
   FiatAccountSchema,
   FiatAccountType,
@@ -19,11 +20,6 @@ import {
 } from 'src/localCurrency/convert'
 import { navigate } from 'src/navigator/NavigationService'
 import { Screens } from 'src/navigator/Screens'
-import { FiatConnectApiClient, FiatConnectClient } from '@fiatconnect/fiatconnect-sdk'
-import { UnlockableWallet } from '@celo/wallet-base'
-import { getSigningFunction } from 'src/fiatconnect/index'
-import { getWalletAsync } from 'src/web3/contracts'
-import { FIATCONNECT_NETWORK } from 'src/config'
 import { Currency, resolveCurrency } from 'src/utils/currencies'
 
 const strings = {
@@ -178,22 +174,5 @@ export default class FiatConnectQuote extends NormalizedQuote {
 
   getFiatAccountSchemaAllowedValues(): { [key: string]: string[] } {
     return this.quoteResponseFiatAccountSchema.allowedValues
-  }
-
-  async getFiatConnectClient(): Promise<FiatConnectApiClient> {
-    if (this.fiatConnectClient) {
-      return this.fiatConnectClient
-    }
-    const wallet = (await getWalletAsync()) as UnlockableWallet
-    const [account] = wallet.getAccounts()
-    this.fiatConnectClient = new FiatConnectClient(
-      {
-        baseUrl: this.getProviderBaseUrl(),
-        network: FIATCONNECT_NETWORK,
-        accountAddress: account,
-      },
-      getSigningFunction(wallet)
-    )
-    return this.fiatConnectClient
   }
 }
